This id field is now int instead of long (is this expected?)
git-svn-id: https://svn.apache.org/repos/asf/incubator/empire-db/trunk@1180840 13f79535-47bb-0310-9956-ffa450edef68

+++ empire-db/src/test/java/org/apache/empire/db/PreparedStatementTest.java
@@ -54,7 +54,7 @@ public class PreparedStatementTest{
         department.setValue(db.DEPARTMENT.BUSINESS_UNIT, "test");
         department.update(conn);
 
-        long id = department.getInt(db.DEPARTMENT.ID);
+        int id = department.getInt(db.DEPARTMENT.ID);
         assertTrue("Department add failed", id > 0);
